feat: 프로필 다이얼로그 수정 및 인터페이스 구현
