+++ presentation/src/main/java/com/example/presentation/view/login/view/OnboardingActivity.kt
@@ -6,7 +6,7 @@ import com.example.presentation.R
 import com.example.presentation.adapter.onboarding.OnboardingAdapter
 import com.example.presentation.base.BaseActivity
 import com.example.presentation.databinding.ActivityOnboardingBinding
-import com.example.presentation.util.PhotoSelectDialog
+import com.example.presentation.view.gallery.view.PhotoSelectDialog
 import com.example.presentation.view.login.viewModel.OnboardingViewModel
 import com.example.presentation.view.main.MainActivity
 import dagger.hilt.android.AndroidEntryPoint
