+++ presentation/src/main/res/layout/activity_gallery.xml
@@ -6,9 +6,49 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_height="match_parent"
         android:layout_width="match_parent">
+        <Toolbar
+            android:id="@+id/tb_gallery"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            android:contentInsetStart="0dp">
+            <ImageButton
+                android:id="@+id/ib_back_up"
+                android:layout_width="@dimen/global_button_size"
+                android:layout_height="@dimen/global_button_size"
+                android:background="@android:color/transparent"
+                android:src="@drawable/ic_arrow"/>
+            <TextView
+                android:id="@+id/tv_gallery_title"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="@string/gallery_all_photo"
+                android:layout_gravity="center"
+                style="@style/title_bold"
+                android:textColor="@color/color_black"/>
+
+            <Button
+                android:id="@+id/ib_next_upload"
+                android:layout_width="@dimen/global_button_size"
+                android:layout_height="@dimen/global_button_size"
+                android:layout_gravity="end"
+                android:background="@android:color/transparent"
+                android:text="업로드"
+                android:visibility="gone"
+                android:padding="0dp"/>
+        </Toolbar>
        <androidx.fragment.app.FragmentContainerView
            android:id="@+id/fcw_gallery"
            android:layout_width="match_parent"
-           android:layout_height="match_parent"/>
+           android:layout_height="0dp"
+           app:layout_constraintTop_toBottomOf="@id/tb_gallery"
+           app:layout_constraintBottom_toBottomOf="parent"
+           app:layout_constraintStart_toStartOf="parent"
+           app:layout_constraintEnd_toEndOf="parent"
+           android:name="androidx.navigation.fragment.NavHostFragment"
+           app:defaultNavHost="true"
+           app:navGraph="@navigation/nav_gallery_graph" />
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
