+++ presentation/src/main/res/layout/fragment_gallery_list.xml
@@ -6,35 +6,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_height="match_parent"
         android:layout_width="match_parent">
-        <Toolbar
-            android:id="@+id/tb_gallery"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            android:contentInsetStart="0dp">
-            <ImageButton
-                android:layout_width="@dimen/global_button_size"
-                android:layout_height="@dimen/global_button_size"
-                android:background="@android:color/transparent"
-                android:src="@drawable/ic_arrow"/>
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="@string/gallery_all_photo"
-                android:layout_gravity="center"
-                style="@style/title_bold"
-                android:textColor="@color/color_black"/>
 
-            <Button
-                android:layout_width="@dimen/global_button_size"
-                android:layout_height="@dimen/global_button_size"
-                android:layout_gravity="end"
-                android:background="@android:color/transparent"
-                android:text="다음"
-                android:padding="0dp"/>
-        </Toolbar>
         <!--        <com.takusemba.cropme.CropLayout-->
         <!--            android:id="@+id/cl_cropper"-->
         <!--            android:layout_width="match_parent"-->
@@ -53,7 +25,7 @@
             android:id="@+id/rv_gallery"
             android:layout_width="match_parent"
             android:layout_height="0dp"
-            app:layout_constraintTop_toBottomOf="@id/tb_gallery"
+            app:layout_constraintTop_toTopOf="parent"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
